separated FeedbackCard from FeedbackPopup
